# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    AgentRole:
      type: string
      enum: [ROUTA, CRAFTER, GATE, DEVELOPER]
...
    TaskStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, REVIEW_REQUIRED, COMPLETED, NEEDS_FIX, BLOCKED, CANCELLED]
    VerificationVerdict:
      type: string
      enum: [APPROVED, NOT_APPROVED, BLOCKED]
...
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
...
    Task:
      type: object
      required: [id, title, objective, status, dependencies, workspaceId, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        objective:
          type: string
        scope:
          type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
        verificationCommands:
          type: array
          items:
            type: string
        assignedTo:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        boardId:
          type: string
        columnId:
          type: string
        position:
          type: integer
        priority:
          $ref: "#/components/schemas/TaskPriority"
        labels:
          type: array
          items:
            type: string
        assignee:
          type: string
        assignedProvider:
          type: string
        assignedRole:
          $ref: "#/components/schemas/AgentRole"
        assignedSpecialistId:
          type: string
        assignedSpecialistName:
          type: string
        triggerSessionId:
          type: string
        githubId:
          type: string
        githubNumber:
          type: integer
        githubUrl:
          type: string
        githubRepo:
          type: string
        githubState:
          type: string
        githubSyncedAt:
          type: string
          format: date-time
        lastSyncError:
          type: string
        dependencies:
          type: array
          items:
            type: string
        parallelGroup:
          type: string
        workspaceId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completionSummary:
          type: string
        verificationVerdict:
          $ref: "#/components/schemas/VerificationVerdict"
        verificationReport:
          type: string
    KanbanColumnAutomation:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        providerId:
          type: string
        role:
          type: string
        specialistId:
          type: string
        specialistName:
          type: string
        transitionType:
          type: string
          enum: [entry, exit, both]
        requiredArtifacts:
          type: array
          items:
            type: string
            enum: [screenshot, test_results, code_diff]
        autoAdvanceOnSuccess:
          type: boolean
    KanbanColumn:
      type: object
      required: [id, name, position, stage]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        position:
          type: integer
        stage:
          type: string
        automation:
          $ref: "#/components/schemas/KanbanColumnAutomation"
    KanbanBoard:
      type: object
      required: [id, workspaceId, name, isDefault, columns, createdAt, updatedAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        isDefault:
          type: boolean
        columns:
          type: array
          items:
            $ref: "#/components/schemas/KanbanColumn"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
paths:
...
  /api/kanban/boards:
    get:
      operationId: listKanbanBoards
      summary: List Kanban boards for a workspace
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: string
            default: "default"
      responses:
        "200":
          description: Board list
          content:
            application/json:
              schema:
                type: object
                properties:
                  boards:
                    type: array
                    items:
                      $ref: "#/components/schemas/KanbanBoard"
    post:
      operationId: createKanbanBoard
      summary: Create a Kanban board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceId, name]
              properties:
                workspaceId:
                  type: string
                name:
                  type: string
                isDefault:
                  type: boolean
      responses:
        "201":
          description: Board created
          content:
            application/json:
              schema:
                type: object
                properties:
                  board:
                    $ref: "#/components/schemas/KanbanBoard"
  /api/kanban/boards/{boardId}:
    get:
      operationId: getKanbanBoard
      summary: Get a Kanban board by ID
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Board details
          content:
            application/json:
              schema:
                type: object
                properties:
                  board:
                    $ref: "#/components/schemas/KanbanBoard"
        "404":
          description: Board not found
    patch:
      operationId: updateKanbanBoard
      summary: Update a Kanban board
      parameters:
        - name: boardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                columns:
                  type: array
                  items:
                    $ref: "#/components/schemas/KanbanColumn"
                isDefault:
                  type: boolean
      responses:
        "200":
          description: Board updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  board:
                    $ref: "#/components/schemas/KanbanBoard"
        "404":
          description: Board not found
  /api/kanban/decompose:
    post:
      operationId: decomposeKanbanTasks
      summary: Decompose natural language input into multiple Kanban tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [boardId, workspaceId, tasks]
              properties:
                boardId:
                  type: string
                workspaceId:
                  type: string
                tasks:
                  type: array
                  items:
                    type: object
                    required: [title]
                    properties:
                      title:
                        type: string
                      description:
                        type: string
                      priority:
                        type: string
                        enum: [low, medium, high, urgent]
                      labels:
                        type: array
                        items:
                          type: string
                columnId:
                  type: string
                  default: backlog
      responses:
        "200":
          description: Tasks created
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  cards:
                    type: array
                    items:
                      $ref: "#/components/schemas/Task"

  # ── Notes ──
...
  /api/kanban/events:
    get:
      operationId: streamKanbanEvents
      summary: Stream kanban workspace events over SSE
      parameters:
        - name: workspaceId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: SSE stream of kanban workspace events
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Specialists ──
...